WoT restructure A1: geometry/ -> domMeasure/ (git-mv + import + eslint couplings)
