+++ handsontable/src/3rdparty/walkontable/src/domMeasure/geometryReader.ts
