+++ handsontable/src/3rdparty/walkontable/src/domMeasure/liveGeometryReader.ts
